pyproject.toml
[project]
name = "recipie-upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "openai>=1.70.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
]
